feat(specualtive-user-aggregator: adds an example using turn.eager_end events from ink
