--- cartesia-studypal/pcc-deploy.toml
@@ -1,5 +1,0 @@
-agent_name = "studypal"
-secret_set = "studypal-secrets"
-
-[scaling]
-	min_agents = 1
