+++ speculative-user-aggregator/pyproject.toml
@@ -1,13 +1,11 @@
 [project]
-name = "cartesia-studypal"
+name = "speculative-user-aggregator"
 version = "0.1.0"
-description = "Cartesia study pal example"
+description = "An example voice agent that uses Ink's turn.eager_end events"
 requires-python = ">=3.11,<4"
 dependencies = [
-    "pipecat-ai[anthropic,google,daily,webrtc,websocket,cartesia,openai,runner]>=1.3.0,<2",
+    "pipecat-ai[daily,webrtc,websocket,cartesia,openai,runner]>=1.3.0,<2",
     "pipecatcloud>=0.4.4,<1",
-    "pypdf>=6,<7",
-    "tiktoken>=0.13.0,<1"
 ]
 
 [dependency-groups]
